pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agent-hive"
version = "0.1.0"
description = "Unified multi-agent system built on LangGraph"
authors = [
    {name = "AgentHive Team", email = "hello@ruac.tech"},
]
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "langchain>=0.1.0",
    "langchain-openai>=0.0.5",
    "langchain-anthropic>=0.1.0",
    "langchain-google-genai>=0.0.6",
    "langgraph>=0.0.40",
    "pydantic>=2.5.0",
    "redis>=5.0.0",
    "python-multipart>=0.0.6",
    "aiofiles>=23.2.1",
    "python-dotenv>=1.0.0",
    "structlog>=23.2.0",
    "pillow>=10.0.0",
    "pypdf2>=3.0.0",
]
requires-python = ">=3.11"

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.25.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.black]
line-length = 88
target-version = ['py311']

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true